lab3/Lab3_Bondarenko_Kryhin/pyproject.toml
[project]
name = "lab3"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
